+++ lib/lhs/concerns/item/validation.rb
@@ -16,7 +16,7 @@ class LHS::Item < LHS::Proxy
       run_validation!(record, options, url, params)
       true
     rescue LHC::Error => e
-      self.errors = LHS::Errors::Base.new(e.response, record)
+      self.errors = LHS::Problems::Errors.new(e.response, record)
       false
     end
     alias validate valid?
