+++ lib/lhs/concerns/proxy/accessors.rb
@@ -67,12 +67,14 @@ class LHS::Proxy
       wrap_return(collection, record, name)
     end
 
-    # Wraps with record and adds nested errors to data,
+    # Wraps with record and adds nested errors/warnings to data,
     # if errors are existing
-    def wrap_return(value, record, name)
+    def wrap_return(value, record, name, args = nil)
+      name = args.first if name == :[]
       return value unless worth_wrapping?(value)
       data = value.is_a?(LHS::Data) || value.is_a?(LHS::Record) ? value : LHS::Data.new(value, _data)
-      data.errors = LHS::Errors::Nested.new(errors, name) if errors
+      data.errors = LHS::Problems::Nested::Errors.new(errors, name) if errors.any?
+      data.warnings = LHS::Problems::Nested::Warnings.new(warnings, name) if warnings.any?
       return record.new(data) if record && !value.is_a?(LHS::Record)
       data
     end
