+++ lib/lhs/problems/nested/errors.rb
@@ -1,0 +1,14 @@
+module LHS::Problems
+  module Nested
+    class Errors < LHS::Problems::Errors
+      include LHS::Problems::Nested::Base
+
+      def initialize(errors, scope)
+        @raw = errors
+        @messages = nest(errors.messages, scope)
+        @message = errors.message
+        @scope = scope
+      end
+    end
+  end
+end
