+++ lib/lhs/problems/nested/warnings.rb
@@ -1,0 +1,13 @@
+module LHS::Problems
+  module Nested
+    class Warnings < LHS::Problems::Warnings
+      include LHS::Problems::Nested::Base
+
+      def initialize(warnings, scope)
+        @raw = warnings.raw
+        @messages = nest(warnings.messages, scope)
+        @scope = scope
+      end
+    end
+  end
+end
